Added DORY_network/ dir for reproducibility

+++ DORY_network/src/pulp_nn_conv_HoWo_parallel.c
@@ -33,8 +33,8 @@ void __attribute__ ((noinline)) pulp_nn_conv_HoWo_parallel(
   const int8_t *  bias,
   uint8_t *       pOutBuffer,
   const int8_t *  pWeight,
-  int32_t *       k,
-  int32_t *       lambda,
+  int64_t *       k,
+  int64_t *       lambda,
   const uint16_t  out_mult,
   const uint16_t  out_shift,
   const uint16_t  dim_in_x,
@@ -52,8 +52,8 @@ void __attribute__ ((noinline)) pulp_nn_conv_HoWo_parallel(
   const uint16_t  stride_x,
   const uint16_t  stride_y,
   int             flag_relu,
-  int             flag_batch_norm)
-{
+  int             flag_batch_norm
+) {
 
   int core_id = pi_core_id();
   uint8_t * pIm2ColBase = pIm2ColBuffer + (2 * core_id * ch_in * dim_kernel_x * dim_kernel_y);
@@ -209,8 +209,8 @@ void __attribute__ ((noinline)) pulp_nn_conv_HoWo_parallel(
     {
       const int8_t *pA = pWeight;
       int       i;
-      int32_t * k1 = k;
-      int32_t * lambda1 = lambda;
+      int64_t * k1 = k;
+      int64_t * lambda1 = lambda;
       for (i = 0; i < ch_out; i++)
       {
         /* include the accumulation buffer in sum computation (probably doesn't work). Maybe the reloading partial result is needed as well as internally at mat mul function. */
